# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OIF Aggregator API",
    "description": "Open Intents Framework (OIF) Aggregator provides a unified API for cross-chain transaction aggregation. This service connects to multiple solvers to find the best execution paths for user intents across different blockchain networks.",
    "contact": {
      "name": "OIF Team",
      "url": "https://github.com/openintentsframework/"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/openintentsframework/oif-aggregator/blob/main/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Submit a new order\nPOST /v1/orders - Submit an order",
        "operationId": "post_orders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quotes": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Get quotes for a swap request using standard OIF format\nPOST /v1/quotes - Get quotes",
        "operationId": "post_quotes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Quotes aggregated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/solvers": {
      "get": {
        "tags": [
          "solvers"
        ],
        "summary": "GET /v1/solvers - List all solvers",
        "operationId": "get_solvers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Items per page (1-100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 25
          }
        ],
        "responses": {
          "200": {
            "description": "List of solvers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolversResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/solvers/{id}": {
      "get": {
        "tags": [
          "solvers"
        ],
        "summary": "GET /v1/solvers/{id} - Get solver by id",
        "operationId": "get_solver_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Solver ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solver details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetAmount": {
        "type": "object",
        "description": "Asset amount representation using ERC-7930 interoperable address format.\n\nThis is a fundamental model used across all OIF versions.",
        "required": [
          "asset"
        ],
        "properties": {
          "amount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/U256",
                "description": "Amount as a big integer"
              }
            ]
          },
          "asset": {
            "$ref": "#/components/schemas/InteropAddress",
            "description": "Asset address in ERC-7930 interoperable format"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response format shared by handlers",
        "required": [
          "error",
          "message",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "OrderRequest": {
        "type": "object",
        "description": "API request body for submitting orders - flexible design for multi-adapter support",
        "required": [
          "quoteResponse",
          "signature"
        ],
        "properties": {
          "metadata": {
            "description": "Adapter-specific metadata that can store order data, sponsor info, and other custom data\nThis allows flexibility for different adapters to include the specific information they need"
          },
          "originSubmission": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OriginSubmission"
              }
            ]
          },
          "quoteResponse": {
            "$ref": "#/components/schemas/QuoteResponse",
            "description": "Quote data"
          },
          "signature": {
            "type": "string",
            "description": "User's signature for authorization"
          }
        },
        "additionalProperties": false,
        "example": {
          "metadata": {
            "order": "0xfedcba0987654321...",
            "sponsor": "0x70997970c51812dc3a010c7d01b50e0d17dc79c8"
          },
          "originSubmission": {
            "mode": "user",
            "schemes": [
              "erc4337",
              "permit2"
            ]
          },
          "quoteResponse": {
            "eta": 30,
            "failureHandling": "refund-automatic",
            "integrityChecksum": "hmac-sha256:a1b2c3d4e5f6...",
            "metadata": {
              "aggregator": "metadata"
            },
            "order": {
              "payload": {
                "domain": {
                  "chainId": 1,
                  "name": "TestDomain",
                  "version": "1"
                },
                "message": {
                  "amount": "1000000000000000000",
                  "inputAsset": "0x00010000027a69145FbDB2315678afecb367f032d93F642f64180aa3",
                  "orderType": "swap",
                  "outputAsset": "0x00010000027a6a145FbDB2315678afecb367f032d93F642f64180aa3"
                },
                "primaryType": "Order",
                "signatureType": "eip712",
                "types": {}
              },
              "type": "oif-escrow-v0"
            },
            "partialFill": false,
            "preview": {
              "inputs": [
                {
                  "amount": "1000000000000000000",
                  "asset": "0x00010000027a69145FbDB2315678afecb367f032d93F642f64180aa3",
                  "user": "0x00010000027a691470997970C51812dc3A010C7d01b50e0d17dc79C8"
                }
              ],
              "outputs": [
                {
                  "amount": "1000000",
                  "asset": "0x00010000027a6a145FbDB2315678afecb367f032d93F642f64180aa3",
                  "receiver": "0x00010000027a6a143C44CdDdB6a900fa2b585dd299e03d12FA4293BC"
                }
              ]
            },
            "provider": "Example Solver v1.0",
            "quoteId": "6a22e92f-3e5d-4f05-ab5f-007b01e58b21",
            "solverId": "example-solver",
            "validUntil": 1756457492
          },
          "signature": "0x1234567890abcdef..."
        }
      },
      "OrderResponse": {
        "type": "object",
        "description": "Response body for /v1/orders endpoint (order submission)",
        "required": [
          "orderId",
          "status",
          "createdAt",
          "updatedAt",
          "inputAmounts",
          "outputAmounts",
          "settlement"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the response was created"
          },
          "fillTransaction": {
            "description": "Fill transaction information"
          },
          "inputAmounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetAmount"
            },
            "description": "Input amount"
          },
          "orderId": {
            "type": "string",
            "description": "The order ID"
          },
          "outputAmounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetAmount"
            },
            "description": "Output amount"
          },
          "quoteId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Quote ID"
          },
          "settlement": {
            "$ref": "#/components/schemas/Settlement",
            "description": "Settlement information"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "description": "Current status of the order"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the response was last updated"
          }
        },
        "additionalProperties": false
      },
...
      "OriginSubmission": {
        "type": "object",
        "description": "Origin submission preference\n\nCore origin submission mechanism that is stable across versions.",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/OriginMode"
          },
          "schemes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/AuthScheme"
            }
          }
        }
      },
...
      "QuoteRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GetQuoteRequest"
          },
          {
            "type": "object",
            "properties": {
              "metadata": {
                "description": "Optional metadata for custom adapter use\nThis allows clients to include adapter-specific information in quote requests"
              },
              "solverOptions": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/SolverOptions",
                    "description": "Solver options"
                  }
                ]
              }
            }
          }
        ],
        "description": "API request body for /v1/quotes endpoint",
        "example": {
          "intent": {
            "inputs": [
              {
                "amount": "1000000000000000000",
                "asset": "0x00010000027a69145FbDB2315678afecb367f032d93F642f64180aa3",
                "user": "0x00010000027a691470997970C51812dc3A010C7d01b50e0d17dc79C8"
              }
            ],
            "intentType": "oif-swap",
            "minValidUntil": 600,
            "outputs": [
              {
                "amount": "1000000",
                "asset": "0x00010000027a6a145FbDB2315678afecb367f032d93F642f64180aa3",
                "receiver": "0x00010000027a6a143C44CdDdB6a900fa2b585dd299e03d12FA4293BC"
              }
            ],
            "partialFill": false,
            "preference": "speed",
            "swapType": "exact-input"
          },
          "solverOptions": {
            "solverTimeout": 2000,
            "timeout": 4000
          }
        }
      },
...
      "QuotesResponse": {
        "type": "object",
        "description": "Collection of quotes response for API endpoints",
        "required": [
          "quotes",
          "totalQuotes"
        ],
        "properties": {
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AggregationMetadata",
                "description": "Optional metadata about the aggregation process"
              }
            ]
          },
          "quotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteResponse"
            },
            "description": "List of quotes"
          },
          "totalQuotes": {
            "type": "integer",
            "description": "Total number of quotes",
            "minimum": 0
          }
        },
        "additionalProperties": false,
        "example": {
          "metadata": {
            "earlyTermination": false,
            "globalTimeoutMs": 10000,
            "minQuotesRequired": 30,
            "solverSelectionMode": "all",
            "solverTimeoutMs": 5000,
            "solversQueried": 2,
            "solversRespondedError": 0,
            "solversRespondedSuccess": 1,
            "solversTimedOut": 1,
            "totalDurationMs": 1500,
            "totalSolversAvailable": 2
          },
          "quotes": [
            {
              "eta": 30,
              "failureHandling": "refund-automatic",
              "integrityChecksum": "hmac-sha256:a1b2c3d4e5f6...",
              "metadata": {
                "aggregator": "metadata"
              },
              "order": {
                "payload": {
                  "domain": {
                    "chainId": 1,
                    "name": "TestDomain",
                    "version": "1"
                  },
                  "message": {
                    "amount": "1000000000000000000",
                    "inputAsset": "0x00010000027a69145FbDB2315678afecb367f032d93F642f64180aa3",
                    "orderType": "swap",
                    "outputAsset": "0x00010000027a6a145FbDB2315678afecb367f032d93F642f64180aa3"
                  },
                  "primaryType": "Order",
                  "signatureType": "eip712",
                  "types": {}
                },
                "type": "oif-escrow-v0"
              },
              "partialFill": false,
              "preview": {
                "inputs": [
                  {
                    "amount": "1000000000000000000",
                    "asset": "0x00010000027a69145FbDB2315678afecb367f032d93F642f64180aa3",
                    "user": "0x00010000027a691470997970C51812dc3A010C7d01b50e0d17dc79C8"
                  }
                ],
                "outputs": [
                  {
                    "amount": "1000000",
                    "asset": "0x00010000027a6a145FbDB2315678afecb367f032d93F642f64180aa3",
                    "receiver": "0x00010000027a6a143C44CdDdB6a900fa2b585dd299e03d12FA4293BC"
                  }
                ]
              },
              "provider": "Example Solver v1.0",
              "quoteId": "6a22e92f-3e5d-4f05-ab5f-007b01e58b21",
              "solverId": "example-solver",
              "validUntil": 1756457492
            }
          ],
          "totalQuotes": 1
        }
      },
...
      "SolverResponse": {
        "type": "object",
        "description": "Response format for individual solvers in API",
        "required": [
          "solverId",
          "adapterId",
          "endpoint",
          "status",
          "supportedAssets",
          "createdAt"
        ],
        "properties": {
          "adapterId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoint": {
            "type": "string"
          },
          "lastSeen": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "solverId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SolverStatus"
          },
          "supportedAssets": {
            "$ref": "#/components/schemas/SupportedAssetsResponse"
          }
        },
        "additionalProperties": false,
        "example": {
          "adapterId": "oif-adapter-v1",
          "createdAt": "2025-08-28T00:00:00Z",
          "description": "An example solver for cross-chain swaps",
          "endpoint": "https://api.example-solver.com",
          "lastSeen": "2025-08-28T16:04:52Z",
          "name": "Example DeFi Solver",
          "solverId": "example-solver",
          "status": "active",
          "supportedAssets": {
            "assets": [
              {
                "address": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
                "chainId": 1,
                "decimals": 18,
                "name": "Wrapped Ether",
                "symbol": "WETH"
              }
            ],
            "source": "autoDiscovered",
            "type": "assets"
          }
        }
      },
...
      "SolverStatus": {
        "type": "string",
        "description": "Solver operational status",
        "enum": [
          "active",
          "disabled"
        ]
      },
      "SolversResponse": {
        "type": "object",
        "description": "Collection of solvers response for API endpoints",
        "required": [
          "solvers",
          "totalSolvers"
        ],
        "properties": {
          "solvers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SolverResponse"
            }
          },
          "totalSolvers": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false,
        "example": {
          "solvers": [
            {
              "adapterId": "oif-adapter-v1",
              "createdAt": "2025-08-28T00:00:00Z",
              "description": "An example solver for cross-chain swaps",
              "endpoint": "https://api.example-solver.com",
              "lastSeen": "2025-08-28T16:04:52Z",
              "name": "Example DeFi Solver",
              "solverId": "example-solver",
              "status": "active",
              "supportedAssets": {
                "assets": [
                  {
                    "address": "0xA0b86a33E6441E7C81F7C93451777f5F4dE78e86",
                    "chainId": 1,
                    "decimals": 6,
                    "name": "USD Coin",
                    "symbol": "USDC"
                  }
                ],
                "source": "autoDiscovered",
                "type": "assets"
              }
            },
            {
              "adapterId": "uniswap-adapter-v1",
              "createdAt": "2025-08-28T00:00:00Z",
              "description": "Uniswap V3 liquidity pools solver",
              "endpoint": "https://api.uniswap.solver.com",
              "lastSeen": "2025-08-28T16:04:50Z",
              "name": "Uniswap V3 Solver",
              "solverId": "uniswap-solver",
              "status": "active",
              "supportedAssets": {
                "routes": [
                  {
                    "destinationChainId": 137,
                    "destinationTokenAddress": "0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619",
                    "destinationTokenSymbol": "WETH",
                    "originChainId": 1,
                    "originTokenAddress": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
                    "originTokenSymbol": "WETH"
                  }
                ],
                "source": "config",
                "type": "routes"
              }
            }
          ],
          "totalSolvers": 2
        }
      },
...
    }
  },
...
}